requirements.txt
crewai
crewai-tools
groq
langchain-groq
tavily-python
streamlit
python-dotenv
pydantic